pyproject.toml
[project]
name = "fastapi-observable"
version = "0.1.0"
description = "Enterprise-grade FastAPI application with observability (metrics, traces, logs, profiling)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.2",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-exporter-prometheus>=0.60b1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "pydantic>=2.0",
    "pydantic-settings>=2.12.0",
    "python-json-logger>=4.0.0",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "locust>=2.42.6",
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]

[tool.ruff]
line-length = 99
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (ruff format handles this)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]

